chore: Removed callback-specific wrapper, and restored ESM async/CJS sync code paths
